Split ASP.NET Core dependencies into separate project.
