+++ src/LtiAdvantage.AspNetCore/AssignmentGradeServices/AddLineItemRequest.cs
@@ -1,4 +1,6 @@
-﻿namespace LtiAdvantage.AssignmentGradeServices
+﻿using LtiAdvantage.AssignmentGradeServices;
+
+namespace LtiAdvantage.AspNetCore.AssignmentGradeServices
 {
     /// <summary>
     /// Represents a create item request request.
