+++ src/LtiAdvantage.AspNetCore/AssignmentGradeServices/GetResultsRequest.cs
@@ -1,4 +1,4 @@
-﻿namespace LtiAdvantage.AssignmentGradeServices
+﻿namespace LtiAdvantage.AspNetCore.AssignmentGradeServices
 {
     /// <summary>
     /// Represents a GetResults request.
