+++ src/LtiAdvantage.AspNetCore/AssignmentGradeServices/ILineItemsController.cs
@@ -1,0 +1,9 @@
+﻿namespace LtiAdvantage.AspNetCore.AssignmentGradeServices
+{
+    /// <summary>
+    /// Marker for ILogger dependency injection.
+    /// </summary>
+    public interface ILineItemsController
+    {
+    }
+}
