+++ src/LtiAdvantage.AspNetCore/AssignmentGradeServices/IResultsController.cs
@@ -1,9 +1,9 @@
-﻿namespace LtiAdvantage.AssignmentGradeServices
+﻿namespace LtiAdvantage.AspNetCore.AssignmentGradeServices
 {
     /// <summary>
     /// Marker for ILogger dependency injection.
     /// </summary>
-    public interface ILineItemsController
+    public interface IResultsController
     {
     }
 }
