+++ src/LtiAdvantage.AspNetCore/README.md
@@ -1,0 +1,6 @@
+# LtiAdvantage.AspNetCore
+
+This library provides ASP.NET Core controller implementations to support platform development.
+It targets `netstandard2.0` for ASP.NET Core 2.1 and 2.2 and `netcoreapp3.1` for ASP.NET Core 3.1.
+It uses BouncyCastle.NetCore to read and write PEM formatted keys
+for compatibility with the [IMS LTI Reference Implementation](https://github.com/IMSGlobal/lti-reference-implementation).
