+++ src/LtiAdvantage.AspNetCore/Utilities/AuthorizationScopePolicyProvider.cs
@@ -3,7 +3,7 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 
-namespace LtiAdvantage.Utilities
+namespace LtiAdvantage.AspNetCore.Utilities
 {
     /// <inheritdoc />
     /// <summary>
