--- src/LtiAdvantage/AssignmentGradeServices/IResultsController.cs
@@ -1,9 +1,0 @@
-﻿namespace LtiAdvantage.AssignmentGradeServices
-{
-    /// <summary>
-    /// Marker for ILogger dependency injection.
-    /// </summary>
-    public interface IResultsController
-    {
-    }
-}
