+++ src/LtiAdvantage/README.md
@@ -7,11 +7,6 @@ LtiAdvantage is a library to ease the work of creating an LTI Advantage platform
   - [Names and Role Provisioning Services v2.0](https://www.imsglobal.org/spec/lti-nrps/v2p0)
 - [IMS Security Framework v1.0](https://www.imsglobal.org/spec/security/v1p0/)
 
-## ASP.NET Core
-
-This library targets `netstandard2.0` for ASP.NET Core 2.1 and 2.2 and `netcoreapp3.1` for ASP.NET Core 3.1.
-It uses BouncyCastle.NetCore to read and write PEM formatted keys
-for compatibility with the [IMS LTI Reference Implementation](https://github.com/IMSGlobal/lti-reference-implementation).
 
 ## Getting Started
