+++ test/LtiAdvantage.IntegrationTests/Controllers/ScoresController.cs
@@ -1,4 +1,5 @@
 ﻿using System.Threading.Tasks;
+using LtiAdvantage.AspNetCore.AssignmentGradeServices;
 using LtiAdvantage.AssignmentGradeServices;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc;
