+++ test/LtiAdvantage.IntegrationTests/LtiAdvantage.IntegrationTests.csproj
@@ -29,6 +29,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\..\src\LtiAdvantage.AspNetCore\LtiAdvantage.AspNetCore.csproj" />
     <ProjectReference Include="..\..\src\LtiAdvantage\LtiAdvantage.csproj" />
   </ItemGroup>
